BackEnd/requirements.txt




fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.5.0
motor==3.3.2
pydantic==2.4.2
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
pypdf2==3.0.1
pdfplumber==0.10.3
langchain==0.0.350
langchain-openai==0.0.2
python-multipart==0.0.6
cors==1.0.1
pytest==7.4.3